# docs/swagger.yaml
basePath: /
definitions:
...
  api.paymentDetectedReq:
    properties:
      amount_minor:
        description: optional override; if nil, use order.amount_minor
        type: integer
      order_id:
        type: string
      tx_hash:
        type: string
    type: object
  api.paymentDetectedResp:
    properties:
      message:
        type: string
      order_id:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Crypto payment processor API.
  title: OSPay API
  version: "1.0"
paths:
...
  /events/payment-detected:
    post:
      consumes:
      - application/json
      description: Notify the system of an on-chain payment for an order
      parameters:
      - description: Payment info
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/api.paymentDetectedReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.paymentDetectedResp'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.paymentDetectedResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Detect payment event
      tags:
      - events
...
  /reconciliation:
    get:
      description: Returns balance and settlement data for a merchant and asset
      parameters:
      - description: Merchant ID
        in: query
        name: merchant_id
        required: true
        type: string
      - description: Asset symbol
        in: query
        name: asset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get reconciliation data
      tags:
      - reconciliation
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"